Apply VisionForge R3 dual-machine and server repairs
Merged after ZIP re-audit and code-gate repair. Native, Windows Host, Sidecar, Android compile/self-tests, architecture and transport gates were validated. Release TLS SPKI/formal secure-data-plane gates remain external security-material boundaries and are intentionally fail-closed.

+++ dual_machine_runtime/host/windows/tests/dxgi_desktop_capture_tests.cpp
@@ -49,8 +49,8 @@ int main() {
         return 77;
     }
     const vfdual::DesktopCaptureRegion region{
-        (output->width - kCaptureEdge) / 2U,
-        (output->height - kCaptureEdge) / 2U,
+        static_cast<std::int32_t>((output->width - kCaptureEdge) / 2U),
+        static_cast<std::int32_t>((output->height - kCaptureEdge) / 2U),
         kCaptureEdge,
         kCaptureEdge,
     };
